smartest/smartest/requirements.txt
django==4.2.9
django-admin-interface==0.28.5
django-colorfield==0.8.0
django-versatileimagefield==3.0
django-money==3.4.1
django-tinymce==3.5.0
django-import-export==3.1.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-tables2==2.5.3
django-filter==22.1
django-registration-redux==2.13
django-extensions==3.2.1
django-compressor==4.3.1
django-cors-headers==4.3.1
python-decouple==3.8
django-rosetta==0.10.0
django-ckeditor==6.1.0
django-htmlmin==0.11.0
django-weasyprint
djangorestframework
markdown
django-rest-passwordreset
coreapi
psycopg2-binary
djangorestframework-simplejwt
